+++ src/operations.h
@@ -47,8 +47,8 @@ public:
 };
 
 void ScrubString(std::string &input, bool forInteger = false);
-void SeperateString(std::string input, std::vector<std::string> &output, bool subDelimiter = false);
-void SeperateFluidOpString(std::string input, std::vector<std::string> &output);
+void SeparateString(std::string input, std::vector<std::string> &output, bool subDelimiter = false);
+void SeparateFluidOpString(std::string input, std::vector<std::string> &output);
 std::string StitchString(std::string stringOne, std::string stringTwo, bool subDelimiter = false);
 std::string StitchString(std::string stringOne, std::string stringTwo, std::string stringThree, bool subDelimiter = false);
 std::string GetRidOfScriptStatement(std::string input, int position = 1);